test(agents): consolidate agent test suites 10 files -> 5
Merge by layer: agent+space helpers -> test_agent_helpers.py;
create/delete/list command flows -> test_agent_command_flows.py;
catalog loading, launch content, get_definition -> test_agent_catalog.py.
test_agent_logic_properties.py (pure logic) and
test_agent_parameter_store.py (own autouse fixture regime) unchanged.
Identical test set before and after: 135 collected, all pass.
